+ create exceptions package and move clause exception to there
