+++ src/global/exceptions/InvalidClauseException.java
@@ -1,4 +1,4 @@
-package global.tools;
+package global.exceptions;
 
 public class InvalidClauseException extends Exception {
     public InvalidClauseException(String message) {
